--- src/components/providers/MUIProvider.tsx
@@ -1,15 +1,0 @@
-'use client';
-
-import { theme } from '@/theme';
-
-import { CssBaseline, ThemeProvider } from '@mui/material';
-import { AppRouterCacheProvider } from '@mui/material-nextjs/v15-appRouter';
-
-export const MUIProvider = ({ children }: { children: React.ReactNode }) => (
-  <AppRouterCacheProvider>
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      {children}
-    </ThemeProvider>
-  </AppRouterCacheProvider>
-);
